add more setting to jseg calculation

+++ otherMethods/JSEG/run_jseg.m
@@ -6,8 +6,8 @@ github_dir = 'C:\Users\luong_nguyen\Documents\GitHub\HE-tumor-object-segmentatio
 jseg_dir = fullfile(github_dir,'otherMethods','JSEG');
 cd(jseg_dir);
 im_dir = 'Z:\Tiles_512_jpg';
-%output_dir = fullfile('Z:\HEproject\evaluation_results\JSEG','multi_scale');
-output_dir = fullfile('Z:\HEproject\evaluation_results\JSEG','one_scale');
+output_dir = fullfile('Z:\HEproject\evaluation_results\JSEG','multi_scale');
+%output_dir = fullfile('Z:\HEproject\evaluation_results\JSEG','one_scale');
 
 if ~exist(output_dir,'dir')
     mkdir(output_dir);
@@ -29,7 +29,7 @@ if ~exist(fullfile(output_dir,'gif_files'),'dir')
     mkdir(fullfile(output_dir,'gif_files'));
 end
 
-quantize_threshold = 0:25:600;
+quantize_threshold = 25:25:600;
 num_thres = length(quantize_threshold);
 im_list = dir(fullfile(im_dir,'*.jpg'));
 im_list = {im_list.name}';
@@ -43,11 +43,10 @@ for i = 1:num_images
         q_thresh = quantize_threshold(j);
         gif_file = fullfile(output_dir,'gif_files',[im_name '_'  num2str(q_thresh) '.gif']);
         if ~exist(gif_file,'file')
-            %expr = ['segwin -i ', fullfile(im_dir,im_list{i}), ' -t 6 -o ', ...
-            %  fullfile(output_dir,[im_name '.jpg']), ' 0.9 -r9 ', ...
-            %  fullfile(output_dir,[im_name '.gif']), ' -q 255'];
             expr = ['segwin -i ', fullfile(im_dir,im_list{i}), ' -t 6 -r9 ', ...
-                gif_file,' -l 1 -q ' num2str(q_thresh)];
+              gif_file,' -q ' num2str(q_thresh)];
+            %expr = ['segwin -i ', fullfile(im_dir,im_list{i}), ' -t 6 -r9 ', ...
+            %    gif_file,' -l 1 -q ' num2str(q_thresh)];
             out_expr = evalc(['system(' quote expr quote ')']);
         end
     end
@@ -65,16 +64,19 @@ parfor i = 1:num_images
     segs = cell(num_thres,1);
     for j = 1:num_thres
         q_thresh = quantize_threshold(j);
-        bdry_im_fname = fullfile(output_dir,'bdry_im',[im_name, '_' num2str(q_thresh), '_bdry.jpg']);
-        labels = imread(fullfile(output_dir,'gif_files',[im_name '_'  num2str(q_thresh) '.gif']),1);
-        segs{j} = labels+1;
-        if ~exist(bdry_im_fname,'file')
-            edge_map = seg2bdry(labels,'imageSize');
-            % change thickness of edges
-            edge_map = imdilate(edge_map, strel('disk',1));
-            edge_map_im = I.*uint8(repmat(~edge_map,[1 1 3]));
-            imwrite(edge_map_im,fullfile(output_dir,'bdry_im',[im_name, '_' num2str(q_thresh), '_bdry.jpg']));
-            imwrite(label2rgb(labels),fullfile(output_dir,'seg_im',[im_name, '_' num2str(q_thresh), '_seg.jpg']));
+        gif_file = fullfile(output_dir,'gif_files',[im_name '_'  num2str(q_thresh) '.gif']);
+        if exist(gif_file,'file')
+            bdry_im_fname = fullfile(output_dir,'bdry_im',[im_name, '_' num2str(q_thresh), '_bdry.jpg']);
+            labels = imread(fullfile(output_dir,'gif_files',[im_name '_'  num2str(q_thresh) '.gif']),1);
+            segs{j} = labels+1;
+            if ~exist(bdry_im_fname,'file')
+                edge_map = seg2bdry(labels,'imageSize');
+                % change thickness of edges
+                edge_map = imdilate(edge_map, strel('disk',1));
+                edge_map_im = I.*uint8(repmat(~edge_map,[1 1 3]));
+                imwrite(edge_map_im,fullfile(output_dir,'bdry_im',[im_name, '_' num2str(q_thresh), '_bdry.jpg']));
+                imwrite(label2rgb(labels),fullfile(output_dir,'seg_im',[im_name, '_' num2str(q_thresh), '_seg.jpg']));
+            end
         end
     end
     parsave(fullfile(output_dir,'mat_files',[im_name '.mat']),segs);
